Actualizadas las clases intervalo, los atributos modificados a punto

+++ src/intervals/ExactPoint.java
@@ -2,7 +2,7 @@ package intervals;
 
 public class ExactPoint extends Point{
 
-	public ExactPoint(int number) {
+	public ExactPoint(double number) {
 		super(number);
 	}
